python-backend/requirements.txt
flask
flask-cors
openai
python-dotenv
requests
pymongo
python-dotenv
pymysql